# openapi.yaml
openapi: 3.0.1
info:
  title: Replicate models
  description: A plugin that allows the user to run Replicate models from ChatGPT.
  version: 'v1'
servers:
  - url: {APP_HOST}
paths:
  /collections:
    post:
      operationId: getCollections
      summary: get a list of available collections, descriptions and their slugs
      responses:
        "200":
          description: OK
  /collection:
    post:
      operationId: getCollection
      summary: get a list of models in a collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getCollectionRequest'
      responses:
        "200":
          description: OK
  /model:
    post:
      operationId: getModel
      summary: get information about a model and how to run it
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/getModelRequest'
      responses:
        "200":
          description: OK
  /run:
    post:
      operationId: runModel
      summary: run a model on Replicate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/runModelRequest'
      responses:
        "200":
          description: OK

components:
  schemas:
    getCollectionRequest:
      type: object
      properties:
        collection_slug:
          type: string
          description: slug of the collection
          required: true
    getModelResponse:
      type: string
      description: JSON representation of the model and its input parameters, including version number
    getModelRequest:
      type: object
      properties:
        username:
          type: string
          description: owner of the model
          required: true
        model:
          type: string
          description: name of the model
          required: true
    runModelResponse:
      type: string
      description: JSON representation of the output from the model
    runModelRequest:
      type: object
      properties:
        username:
          type: string
          description: owner of the model
          required: true
        model:
          type: string
          description: name of the model
          required: true
        version:
          type: string
          description: the version of the model to use
          required: true
        inputJSON:
          type: string
          description: a stringified JSON object of the input params for the model
          required: true